Refactor LoginAlert and add proptypes
